fix: using existing navbar, included new file for stylying, used svm instead of png
