+++ q-file-share-ui/src/app/dashboard/Dashboard.css
@@ -1,0 +1,66 @@
+/* Navbar styles */
+.navbar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 10px 20px;
+  background-color: #282c34;
+  color: white;
+}
+
+.nav-title {
+  margin: 0;
+}
+
+.logout-button {
+  background-color: #ff4b5c;
+  color: white;
+  border: none;
+  padding: 8px 12px;
+  cursor: pointer;
+  border-radius: 4px;
+}
+
+/* Dashboard container styles */
+.dashboardContainer {
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.cardContainer {
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+  max-width: 1200px;
+  margin-bottom: 20px;
+  margin-top: 150px;
+}
+
+/* Card component styles */
+.card {
+  flex: 1;
+  padding: 20px;
+  margin: 0 10px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  border-radius: 8px;
+  text-align: center;
+  background-color: white;
+  cursor: pointer;
+  transition: transform 0.2s;
+}
+
+.cardLogo {
+  width: 40px;
+  height: 40px;
+  margin-bottom: 10px;
+}
+
+.activity-card {
+  padding: 20px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  border-radius: 8px;
+  background-color: white;
+  width: 100%;
+}
